appModeCheck move api structs
