--- modules/dop/utils/protocol.go
@@ -1,31 +1,0 @@
-// Copyright (c) 2021 Terminus, Inc.
-//
-// This program is free software: you can use, redistribute, and/or modify
-// it under the terms of the GNU Affero General Public License, version 3
-// or later ("AGPL"), as published by the Free Software Foundation.
-//
-// This program is distributed in the hope that it will be useful, but WITHOUT
-// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-// FITNESS FOR A PARTICULAR PURPOSE.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-package utils
-
-import (
-	"os"
-
-	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/pkg/strutil"
-)
-
-// GetProtocol  return protocol
-func GetProtocol() string {
-	protocols := strutil.Split(os.Getenv(string(apistructs.DICE_PROTOCOL)), ",", true)
-	protocol := "https"
-	if len(protocols) > 0 {
-		protocol = protocols[0]
-	}
-	return protocol
-}
